backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-socketio==5.11.0
python-pptx==0.6.23
tenacity==8.2.3
limits==3.12.0
loguru==0.7.2
prometheus-fastapi-instrumentator==7.0.0
